ref(issues): Move stacktraces module to issues folder
