+++ pyproject.toml
@@ -515,6 +515,7 @@ module = [
     "sentry.issues.related.*",
     "sentry.issues.run",
     "sentry.issues.services.*",
+    "sentry.issues.stacktraces.platform",
     "sentry.issues.status_change",
     "sentry.issues.status_change_consumer",
     "sentry.issues.status_change_message",
@@ -616,7 +617,6 @@ module = [
     "sentry.snuba.query_subscriptions.*",
     "sentry.spans.consumers.process_segments.*",
     "sentry.spans.grouping.*",
-    "sentry.stacktraces.platform",
     "sentry.status_checks.*",
     "sentry.tasks.beacon",
     "sentry.tasks.codeowners.*",
