+++ src/sentry/issues/stacktraces/processing.py
@@ -8,9 +8,9 @@ from urllib.parse import urlparse
 
 import sentry_sdk
 
+from sentry.issues.stacktraces.functions import set_in_app, trim_function_name
 from sentry.models.project import Project
 from sentry.models.release import Release
-from sentry.stacktraces.functions import set_in_app, trim_function_name
 from sentry.utils import metrics
 from sentry.utils.cache import cache
 from sentry.utils.hashlib import hash_values
