+++ tests/relay_integration/lang/java/test_plugin.py
@@ -6,9 +6,9 @@ import pytest
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.urls import reverse
 
+from sentry.issues.stacktraces.processing import find_stacktraces_in_data
 from sentry.models.debugfile import ProjectDebugFile
 from sentry.models.files.file import File
-from sentry.stacktraces.processing import find_stacktraces_in_data
 from sentry.testutils.cases import TransactionTestCase
 from sentry.testutils.helpers.datetime import before_now
 from sentry.testutils.relay import RelayStoreHelper
