+++ tests/sentry/grouping/test_categorization.py
@@ -58,7 +58,7 @@ from django.utils.functional import cached_property
 
 from sentry.grouping.api import get_default_grouping_config_dict, load_grouping_config
 from sentry.grouping.strategies.base import StrategyConfiguration
-from sentry.stacktraces.processing import normalize_stacktraces_for_grouping
+from sentry.issues.stacktraces.processing import normalize_stacktraces_for_grouping
 from sentry.testutils.pytest.fixtures import InstaSnapshotter
 from sentry.utils.safe import get_path
