+++ tests/sentry/issues/stacktraces/test_platform.py
@@ -1,6 +1,6 @@
 import pytest
 
-from sentry.stacktraces.platform import get_behavior_family_for_platform
+from sentry.issues.stacktraces.platform import get_behavior_family_for_platform
 
 
 @pytest.mark.parametrize(
