+++ tests/sentry/issues/stacktraces/test_stacktraces.py
@@ -4,7 +4,10 @@ from typing import Any
 
 import pytest
 
-from sentry.stacktraces.processing import find_stacktraces_in_data, get_crash_frame_from_event_data
+from sentry.issues.stacktraces.processing import (
+    find_stacktraces_in_data,
+    get_crash_frame_from_event_data,
+)
 from sentry.testutils.cases import TestCase
